I was able to get the random menu set up and working still need to refactor.
I struggled to find a was to get the random four part menu to work, but I got it to work.  I still need to refactor the code due to it repeating to get the four parts working I want to go back and rework with a loop of some kind I don't know whitch one ill go with yet.